requirements.txt
addict==2.4.0
albumentations==1.3.0
allennlp==2.7.0
captum==0.5.0
click==7.1.2
clip==1.0
dependencies==7.4.0
fast_slic==0.4.0
ffmpeg==1.4
ipykernel==6.16.0
ipython==8.5.0
lib==4.0.0
matplotlib==3.5.3
moviepy==1.0.3
munch==2.5.0
numpy==1.23.1
opencv_python_headless==4.6.0.66
pandas==1.4.3
Pillow==9.2.0
protobuf==4.21.7
PyYAML==6.0
requests==2.28.1
rfconv==0.0.2b20210406
rouge==1.0.0
sacrebleu==1.5.1
scikit_image==0.19.3
scikit_learn==1.1.2
scipy==1.9.0
segmentation_models_pytorch==0.3.0
setuptools==44.0.0
six==1.16.0
skimage==0.0
spacy==3.0.7
spellchecker==0.4
statsmodels==0.13.2
tensorflow==2.10.0
thop==0.1.1.post2209072238
torch==1.9.1
torchfile==0.1.0
torchvideo==0.0.1
tqdm==4.64.0
transformers==4.9.2
wandb==0.12.21
zennit==0.4.6
